+++ agent-allure-spock/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Agent version updated on [5.1.3](https://github.com/reportportal/agent-java-spock/releases/tag/5.1.3), by @HardNorth
 
 ## [5.1.0]
 ### Changed